Avoid reparsing `env_args` continuously
We only need to parse this information once, so calculate it when a
`TargetInfo` is created and then just reuse that from then on out.
