+++ src/cargo/core/compiler/context/compilation_files.rs
@@ -504,9 +504,9 @@ fn compute_metadata<'a, 'cfg>(
     // This helps when the target directory is a shared cache for projects with different cargo configs,
     // or if the user is experimenting with different rustflags manually.
     if unit.mode.is_doc() {
-        cx.bcx.rustdocflags_args(unit).ok().hash(&mut hasher);
+        cx.bcx.rustdocflags_args(unit).hash(&mut hasher);
     } else {
-        cx.bcx.rustflags_args(unit).ok().hash(&mut hasher);
+        cx.bcx.rustflags_args(unit).hash(&mut hasher);
     }
 
     // Artifacts compiled for the host should have a different metadata
